+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The changes not yet present in any release are listed in this section.
 * The whole asus-fan-control API has been remastered.
 * The `-help` option has been changed to the `help` command.
 * The acpi_call module loading is managed by systemd rather than in */etc/modules*.
+* Tab completion has been updated to reflect the API changes.
 
 ### Fixed